requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.9.1+cpu
torchvision==0.24.1+cpu
torchaudio==2.9.1+cpu

pillow
pyyaml

